+++ spec/spec_helper.rb
@@ -1,13 +1,11 @@
 require 'chefspec'
 require 'chefspec/berkshelf'
 
-require_relative '../libraries/selinux_file_helper'
-require_relative '../libraries/selinux_module_helper'
+# Require all our libraries
+Dir['libraries/*.rb'].each { |f| require File.expand_path(f) }
 
 RSpec.configure do |config|
-  config.expect_with :rspec do |c|
-    c.syntax = [:should, :expect]
-  end
+  config.color = true               # Use color in STDOUT
+  config.formatter = :documentation # Use the specified formatter
+  config.log_level = :error         # Avoid deprecation notice SPAM
 end
-
-# EOF